fix tests + extract superclass
